+++ packages/tswebextension/src/lib/mv3/tabs/frame.ts
@@ -1,4 +1,65 @@
-import type { CosmeticResult, MatchingResult } from '@adguard/tsurlfilter';
+import {
+    type CosmeticResult,
+    type MatchingResult,
+    type NetworkRule,
+    type ScriptletData,
+} from '@adguard/tsurlfilter';
+
+/**
+ * Prepared cosmetic result.
+ * This type represents the processed cosmetic data extracted from the initial cosmetic result.
+ */
+type PreparedCosmeticResult = {
+    /**
+     * Script text extracted from the cosmetic result.
+     */
+    scriptText: string;
+
+    /**
+     * A list of scriptlet data extracted from the cosmetic result.
+     */
+    scriptletDataList: ScriptletData[];
+
+    /**
+     * CSS styles extracted from the cosmetic result.
+     */
+    cssText?: string;
+};
+
+/**
+ * Frame constructor properties.
+ */
+type FrameConstructorProps = {
+    /**
+     * Frame url.
+     */
+    url: string;
+
+    /**
+     * Tab id.
+     */
+    tabId: number;
+
+    /**
+     * Frame id.
+     */
+    frameId: number;
+
+    /**
+     * Frame creation time.
+     */
+    timeStamp: number;
+
+    /**
+     * Parent document id.
+     */
+    parentDocumentId?: string;
+
+    /**
+     * Document id.
+     */
+    documentId?: string;
+};
 
 /**
  * Frame context data.
@@ -14,18 +75,43 @@ export class Frame {
     public url: string;
 
     /**
-     * Frame request id.
+     * Tab id.
      */
-    public requestId?: string;
+    public tabId: number;
 
     /**
-     * Frame cosmetic result.
-     * This data is saved in the frame because we need to access it for css injection
-     * after deleting request context data.
+     * Frame id.
+     */
+    public frameId: number;
+
+    /**
+     * Frame creation time.
+     */
+    public timeStamp: number;
+
+    /**
+     * Frame rule. Needed in the case of allowlist rules for the tab.
+     */
+    public frameRule?: NetworkRule;
+
+    /**
+     * Parent document id.
+     */
+    public parentDocumentId?: string;
+
+    /**
+     * The cosmetic result for the frame.
+     * This data is stored in the frame because it is required for logging script and scriptlet rules.
      *
      * @see {@link WebRequestApi.injectCosmetic}
      */
     public cosmeticResult?: CosmeticResult;
+
+    /**
+     * Prepared cosmetic result for the frame.
+     * This data is saved in the frame because it is needed for injecting cosmetic rules into the frames.
+     */
+    public preparedCosmeticResult?: PreparedCosmeticResult;
 
     /**
      * Frame matching result.
@@ -35,13 +121,29 @@ export class Frame {
     public matchingResult?: MatchingResult | null;
 
     /**
-     * Creates frame instance.
-     *
-     * @param url Frame url.
-     * @param requestId Request id.
+     * Unique identifier for the frame.
      */
-    constructor(url: string, requestId?: string) {
+    public documentId?: string;
+
+    /**
+     * Creates frame instance.
+     * @param props Frame constructor properties.
+     */
+    constructor(props: FrameConstructorProps) {
+        const {
+            url,
+            tabId,
+            frameId,
+            timeStamp,
+            parentDocumentId,
+            documentId,
+        } = props;
+
         this.url = url;
-        this.requestId = requestId;
+        this.tabId = tabId;
+        this.frameId = frameId;
+        this.timeStamp = timeStamp;
+        this.parentDocumentId = parentDocumentId;
+        this.documentId = documentId;
     }
 }