+++ packages/tswebextension/test/lib/mv3/background/request/request-context-storage.test.ts
@@ -4,7 +4,6 @@ import {
     RequestContextState,
     RequestContextStorage,
 } from '../../../../../src/lib/mv3/background/request';
-import { FRAME_DELETION_TIMEOUT_MS } from '../../../../../src/lib';
 
 describe('RequestContextStorage', () => {
     let storage: RequestContextStorage;
@@ -44,44 +43,15 @@ describe('RequestContextStorage', () => {
 
     test('should not remove non expired request context by tab and frame id', () => {
         storage.set('requestId4', mockRequestContext);
-        storage.deleteByTabAndFrame(mockRequestContext.tabId, mockRequestContext.frameId);
+        storage.delete(mockRequestContext.requestId);
         const context = storage.get('requestId4');
         expect(context).toBe(mockRequestContext);
-    });
-
-    test('should remove expired request contexts by tab and frame id', () => {
-        mockRequestContext.timestamp = Date.now() - FRAME_DELETION_TIMEOUT_MS - 1000;
-        storage.set('requestId5', mockRequestContext);
-        storage.deleteByTabAndFrame(mockRequestContext.tabId, mockRequestContext.frameId);
-        const context = storage.get('requestId5');
-        expect(context).toBeUndefined();
     });
 
     test('should clear all request contexts', () => {
         storage.set('requestId6', mockRequestContext);
         storage.clear();
         const context = storage.get('requestId6');
-        expect(context).toBeUndefined();
-    });
-
-    test('should get request context by tab and frame id', () => {
-        storage.set('requestId7', mockRequestContext);
-        const context = storage.getByTabAndFrame(mockRequestContext.tabId, mockRequestContext.frameId);
-        expect(context).toEqual(mockRequestContext);
-    });
-
-    test('should handle multiple requests with same tab and frame id', () => {
-        const mockRequestContext2 = { ...mockRequestContext, eventId: 'testEventId2' };
-        storage.set('requestId8', mockRequestContext);
-        storage.set('requestId9', mockRequestContext2);
-        const context = storage.getByTabAndFrame(mockRequestContext.tabId, mockRequestContext.frameId);
-        expect(context).toEqual(mockRequestContext2);
-    });
-
-    test('should not store tab and frame id for non-document requests', () => {
-        mockRequestContext.requestType = RequestType.Image;
-        storage.set('requestId10', mockRequestContext);
-        const context = storage.getByTabAndFrame(mockRequestContext.tabId, mockRequestContext.frameId);
         expect(context).toBeUndefined();
     });
 });