+++ packages/tswebextension/CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version changes -->
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
+## Unreleased
+
+### Fixed
+- MV3 extension cannot apply rules to `about:blank` iframes [AdguardBrowserExtension#2975]
+
+[AdguardBrowserExtension#2975]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2975
+
 ## [2.3.0-alpha.1] - 2024-10-02
 
 ### Changed